requirements.txt
click
numpy
pyyaml
pandas
yfinance
